refactor: create src folder and reorganize file structure
